adds menu.nvgt include, and some vector functions
* Introduce new menu system based on audio form, in menu.nvgt.

* Rename old dynamic menu to bgt_dynamic_menu.nvgt.

* Registered some new vector functions such as cross, length2, distance and distance2 etc.

* datastream.available is now an uint64 rather than int.

+++ src/datastreams.h
@@ -94,7 +94,7 @@ public:
 	inline bool active() {
 		return r || w;
 	}
-	inline int available() {
+	inline unsigned long long available() {
 		return r ? r->available() : -1;
 	}
 	inline std::istream* get_istr() {
